Refactor object integration tests to smaller files.
The integration tests for object operations were too large, in this
change we complete the refactoring into smaller files.
