+++ google/cloud/storage/tests/CMakeLists.txt
@@ -22,10 +22,13 @@ set(storage_client_integration_tests
     curl_resumable_streambuf_integration_test.cc
     curl_resumable_upload_session_integration_test.cc
     curl_streambuf_integration_test.cc
-    object_media_integration_test.cc
+    object_checksum_integration_test.cc
+    object_hash_integration_test.cc
     object_insert_integration_test.cc
     object_integration_test.cc
+    object_media_integration_test.cc
     object_resumable_write_integration_test.cc
+    object_rewrite_integration_test.cc
     service_account_integration_test.cc
     storage_include_test.cc
     thread_integration_test.cc)
