+++ google/cloud/storage/tests/object_integration_test.cc
@@ -375,40 +375,6 @@ TEST_F(ObjectIntegrationTest, ReadNotFound) {
 #endif  // GOOGLE_CLOUD_CPP_HAVE_EXCEPTIONS
 }
 
-TEST_F(ObjectIntegrationTest, Copy) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto source_object_name = MakeRandomObjectName();
-  auto destination_object_name = MakeRandomObjectName();
-
-  std::string expected = LoremIpsum();
-
-  StatusOr<ObjectMetadata> source_meta = client.InsertObject(
-      bucket_name, source_object_name, expected, IfGenerationMatch(0));
-  ASSERT_TRUE(source_meta.ok()) << "status=" << source_meta.status();
-
-  EXPECT_EQ(source_object_name, source_meta->name());
-  EXPECT_EQ(bucket_name, source_meta->bucket());
-
-  StatusOr<ObjectMetadata> meta = client.CopyObject(
-      bucket_name, source_object_name, bucket_name, destination_object_name,
-      WithObjectMetadata(ObjectMetadata().set_content_type("text/plain")));
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-  EXPECT_EQ(destination_object_name, meta->name());
-  EXPECT_EQ(bucket_name, meta->bucket());
-  EXPECT_EQ("text/plain", meta->content_type());
-
-  auto stream = client.ReadObject(bucket_name, destination_object_name);
-  std::string actual(std::istreambuf_iterator<char>{stream}, {});
-  EXPECT_EQ(expected, actual);
-
-  StatusOr<void> status =
-      client.DeleteObject(bucket_name, destination_object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, source_object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
 TEST_F(ObjectIntegrationTest, StreamingWrite) {
   Client client;
   auto bucket_name = ObjectTestEnvironment::bucket_name();
@@ -614,692 +580,6 @@ TEST_F(ObjectIntegrationTest, WriteWithContentType) {
   ASSERT_TRUE(status.ok()) << "status=" << status.status();
 }
 
-TEST_F(ObjectIntegrationTest, CopyPredefinedAclAuthenticatedRead) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-  auto copy_name = MakeRandomObjectName();
-
-  StatusOr<ObjectMetadata> original = client.InsertObject(
-      bucket_name, object_name, LoremIpsum(), IfGenerationMatch(0));
-  ASSERT_TRUE(original.ok()) << "status=" << original.status();
-
-  StatusOr<ObjectMetadata> meta = client.CopyObject(
-      bucket_name, object_name, bucket_name, copy_name, IfGenerationMatch(0),
-      DestinationPredefinedAcl::AuthenticatedRead(), Projection::Full());
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-  EXPECT_LT(0, CountMatchingEntities(meta->acl(),
-                                     ObjectAccessControl()
-                                         .set_entity("allAuthenticatedUsers")
-                                         .set_role("READER")))
-      << *meta;
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, copy_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-TEST_F(ObjectIntegrationTest, CopyPredefinedAclBucketOwnerFullControl) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-  auto copy_name = MakeRandomObjectName();
-
-  StatusOr<BucketMetadata> bucket =
-      client.GetBucketMetadata(bucket_name, Projection::Full());
-  ASSERT_TRUE(bucket.ok()) << "status=" << bucket.status();
-  ASSERT_TRUE(bucket->has_owner());
-  std::string owner = bucket->owner().entity;
-
-  StatusOr<ObjectMetadata> original = client.InsertObject(
-      bucket_name, object_name, LoremIpsum(), IfGenerationMatch(0));
-  ASSERT_TRUE(original.ok()) << "status=" << original.status();
-
-  StatusOr<ObjectMetadata> meta = client.CopyObject(
-      bucket_name, object_name, bucket_name, copy_name, IfGenerationMatch(0),
-      DestinationPredefinedAcl::BucketOwnerFullControl(), Projection::Full());
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-  EXPECT_LT(0, CountMatchingEntities(
-                   meta->acl(),
-                   ObjectAccessControl().set_entity(owner).set_role("OWNER")))
-      << *meta;
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, copy_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-TEST_F(ObjectIntegrationTest, CopyPredefinedAclBucketOwnerRead) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-  auto copy_name = MakeRandomObjectName();
-
-  StatusOr<BucketMetadata> bucket =
-      client.GetBucketMetadata(bucket_name, Projection::Full());
-  ASSERT_TRUE(bucket.ok()) << "status=" << bucket.status();
-  ASSERT_TRUE(bucket->has_owner());
-  std::string owner = bucket->owner().entity;
-
-  StatusOr<ObjectMetadata> original = client.InsertObject(
-      bucket_name, object_name, LoremIpsum(), IfGenerationMatch(0));
-  ASSERT_TRUE(original.ok()) << "status=" << original.status();
-
-  StatusOr<ObjectMetadata> meta = client.CopyObject(
-      bucket_name, object_name, bucket_name, copy_name, IfGenerationMatch(0),
-      DestinationPredefinedAcl::BucketOwnerRead(), Projection::Full());
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-  EXPECT_LT(0, CountMatchingEntities(
-                   meta->acl(),
-                   ObjectAccessControl().set_entity(owner).set_role("READER")))
-      << *meta;
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, copy_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-TEST_F(ObjectIntegrationTest, CopyPredefinedAclPrivate) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-  auto copy_name = MakeRandomObjectName();
-
-  StatusOr<ObjectMetadata> original = client.InsertObject(
-      bucket_name, object_name, LoremIpsum(), IfGenerationMatch(0));
-  ASSERT_TRUE(original.ok()) << "status=" << original.status();
-
-  StatusOr<ObjectMetadata> meta = client.CopyObject(
-      bucket_name, object_name, bucket_name, copy_name, IfGenerationMatch(0),
-      DestinationPredefinedAcl::Private(), Projection::Full());
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-  ASSERT_TRUE(meta->has_owner());
-  EXPECT_LT(
-      0, CountMatchingEntities(meta->acl(), ObjectAccessControl()
-                                               .set_entity(meta->owner().entity)
-                                               .set_role("OWNER")))
-      << *meta;
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, copy_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-TEST_F(ObjectIntegrationTest, CopyPredefinedAclProjectPrivate) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-  auto copy_name = MakeRandomObjectName();
-
-  StatusOr<ObjectMetadata> original = client.InsertObject(
-      bucket_name, object_name, LoremIpsum(), IfGenerationMatch(0));
-  ASSERT_TRUE(original.ok()) << "status=" << original.status();
-
-  StatusOr<ObjectMetadata> meta = client.CopyObject(
-      bucket_name, object_name, bucket_name, copy_name, IfGenerationMatch(0),
-      DestinationPredefinedAcl::ProjectPrivate(), Projection::Full());
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-  ASSERT_TRUE(meta->has_owner());
-  EXPECT_LT(
-      0, CountMatchingEntities(meta->acl(), ObjectAccessControl()
-                                               .set_entity(meta->owner().entity)
-                                               .set_role("OWNER")))
-      << *meta;
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, copy_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-TEST_F(ObjectIntegrationTest, CopyPredefinedAclPublicRead) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-  auto copy_name = MakeRandomObjectName();
-
-  StatusOr<ObjectMetadata> original = client.InsertObject(
-      bucket_name, object_name, LoremIpsum(), IfGenerationMatch(0));
-  ASSERT_TRUE(original.ok()) << "status=" << original.status();
-
-  StatusOr<ObjectMetadata> meta = client.CopyObject(
-      bucket_name, object_name, bucket_name, copy_name, IfGenerationMatch(0),
-      DestinationPredefinedAcl::PublicRead(), Projection::Full());
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-  EXPECT_LT(
-      0, CountMatchingEntities(
-             meta->acl(),
-             ObjectAccessControl().set_entity("allUsers").set_role("READER")))
-      << *meta;
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, copy_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-TEST_F(ObjectIntegrationTest, ComposeSimple) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  // Create the object, but only if it does not exist already.
-  StatusOr<ObjectMetadata> meta = client.InsertObject(
-      bucket_name, object_name, LoremIpsum(), IfGenerationMatch(0));
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-
-  EXPECT_EQ(object_name, meta->name());
-  EXPECT_EQ(bucket_name, meta->bucket());
-
-  // Compose new of object using previously created object
-  auto composed_object_name = MakeRandomObjectName();
-  std::vector<ComposeSourceObject> source_objects = {{object_name},
-                                                     {object_name}};
-  StatusOr<ObjectMetadata> composed_meta = client.ComposeObject(
-      bucket_name, source_objects, composed_object_name,
-      WithObjectMetadata(ObjectMetadata().set_content_type("plain/text")));
-  ASSERT_TRUE(composed_meta.ok()) << "status=" << composed_meta.status();
-  EXPECT_EQ(meta->size() * 2, composed_meta->size());
-
-  StatusOr<void> status =
-      client.DeleteObject(bucket_name, composed_object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-TEST_F(ObjectIntegrationTest, ComposedUsingEncryptedObject) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  std::string content = LoremIpsum();
-  EncryptionKeyData key = MakeEncryptionKeyData();
-
-  // Create the object, but only if it does not exist already.
-  StatusOr<ObjectMetadata> meta =
-      client.InsertObject(bucket_name, object_name, content,
-                          IfGenerationMatch(0), EncryptionKey(key));
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-
-  EXPECT_EQ(object_name, meta->name());
-  EXPECT_EQ(bucket_name, meta->bucket());
-  ASSERT_TRUE(meta->has_customer_encryption());
-  EXPECT_EQ("AES256", meta->customer_encryption().encryption_algorithm);
-  EXPECT_EQ(key.sha256, meta->customer_encryption().key_sha256);
-
-  // Compose new of object using previously created object
-  auto composed_object_name = MakeRandomObjectName();
-  std::vector<ComposeSourceObject> source_objects = {{object_name},
-                                                     {object_name}};
-  StatusOr<ObjectMetadata> composed_meta = client.ComposeObject(
-      bucket_name, source_objects, composed_object_name, EncryptionKey(key));
-  ASSERT_TRUE(composed_meta.ok()) << "status=" << composed_meta.status();
-
-  EXPECT_EQ(meta->size() * 2, composed_meta->size());
-  StatusOr<void> status =
-      client.DeleteObject(bucket_name, composed_object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-TEST_F(ObjectIntegrationTest, RewriteSimple) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto source_name = MakeRandomObjectName();
-
-  // Create the object, but only if it does not exist already.
-  StatusOr<ObjectMetadata> source_meta = client.InsertObject(
-      bucket_name, source_name, LoremIpsum(), IfGenerationMatch(0));
-  ASSERT_TRUE(source_meta.ok()) << "status=" << source_meta.status();
-
-  EXPECT_EQ(source_name, source_meta->name());
-  EXPECT_EQ(bucket_name, source_meta->bucket());
-
-  // Rewrite object into a new object.
-  auto object_name = MakeRandomObjectName();
-  StatusOr<ObjectMetadata> rewritten_meta = client.RewriteObjectBlocking(
-      bucket_name, source_name, bucket_name, object_name);
-  ASSERT_TRUE(rewritten_meta.ok()) << "status=" << rewritten_meta.status();
-
-  EXPECT_EQ(bucket_name, rewritten_meta->bucket());
-  EXPECT_EQ(object_name, rewritten_meta->name());
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, source_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-TEST_F(ObjectIntegrationTest, RewriteEncrypted) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto source_name = MakeRandomObjectName();
-
-  // Create the object, but only if it does not exist already.
-  EncryptionKeyData source_key = MakeEncryptionKeyData();
-  StatusOr<ObjectMetadata> source_meta =
-      client.InsertObject(bucket_name, source_name, LoremIpsum(),
-                          IfGenerationMatch(0), EncryptionKey(source_key));
-  ASSERT_TRUE(source_meta.ok()) << "status=" << source_meta.status();
-
-  EXPECT_EQ(source_name, source_meta->name());
-  EXPECT_EQ(bucket_name, source_meta->bucket());
-
-  // Compose new of object using previously created object
-  auto object_name = MakeRandomObjectName();
-  EncryptionKeyData dest_key = MakeEncryptionKeyData();
-  ObjectRewriter rewriter = client.RewriteObject(
-      bucket_name, source_name, bucket_name, object_name,
-      SourceEncryptionKey(source_key), EncryptionKey(dest_key));
-
-  StatusOr<ObjectMetadata> rewritten_meta = rewriter.Result();
-  ASSERT_TRUE(rewritten_meta.ok()) << "status=" << rewritten_meta.status();
-
-  EXPECT_EQ(bucket_name, rewritten_meta->bucket());
-  EXPECT_EQ(object_name, rewritten_meta->name());
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, source_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-TEST_F(ObjectIntegrationTest, RewriteLarge) {
-  // The testbench always requires multiple iterations to copy this object.
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto source_name = MakeRandomObjectName();
-
-  std::string large_text;
-  long const lines = 8 * 1024 * 1024 / 128;
-  for (long i = 0; i != lines; ++i) {
-    auto line = google::cloud::internal::Sample(generator_, 127,
-                                                "abcdefghijklmnopqrstuvwxyz"
-                                                "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
-                                                "012456789");
-    large_text += line + "\n";
-  }
-
-  StatusOr<ObjectMetadata> source_meta = client.InsertObject(
-      bucket_name, source_name, large_text, IfGenerationMatch(0));
-  ASSERT_TRUE(source_meta.ok()) << "status=" << source_meta.status();
-
-  EXPECT_EQ(source_name, source_meta->name());
-  EXPECT_EQ(bucket_name, source_meta->bucket());
-
-  // Rewrite object into a new object.
-  auto object_name = MakeRandomObjectName();
-  ObjectRewriter writer =
-      client.RewriteObject(bucket_name, source_name, bucket_name, object_name);
-
-  StatusOr<ObjectMetadata> rewritten_meta =
-      writer.ResultWithProgressCallback([](StatusOr<RewriteProgress> const& p) {
-        ASSERT_TRUE(p.ok()) << "progress status=" << p.status();
-        EXPECT_TRUE((p->total_bytes_rewritten < p->object_size) xor p->done)
-            << "p.done=" << p->done << ", p.object_size=" << p->object_size
-            << ", p.total_bytes_rewritten=" << p->total_bytes_rewritten;
-      });
-  ASSERT_TRUE(rewritten_meta.ok()) << "status=" << rewritten_meta.status();
-
-  EXPECT_EQ(bucket_name, rewritten_meta->bucket());
-  EXPECT_EQ(object_name, rewritten_meta->name());
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-  status = client.DeleteObject(bucket_name, source_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that MD5 hashes are computed by default.
-TEST_F(ObjectIntegrationTest, DefaultMD5HashXML) {
-  Client client(ClientOptions()
-                    .set_enable_raw_client_tracing(true)
-                    .set_enable_http_tracing(true));
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  auto backend = std::make_shared<testing_util::CaptureLogLinesBackend>();
-  auto id = LogSink::Instance().AddBackend(backend);
-  StatusOr<ObjectMetadata> insert_meta = client.InsertObject(
-      bucket_name, object_name, LoremIpsum(), IfGenerationMatch(0), Fields(""));
-  ASSERT_TRUE(insert_meta.ok()) << "status=" << insert_meta.status();
-
-  LogSink::Instance().RemoveBackend(id);
-
-  auto count =
-      std::count_if(backend->log_lines.begin(), backend->log_lines.end(),
-                    [](std::string const& line) {
-                      return line.rfind("x-goog-hash: md5=", 0) == 0;
-                    });
-  EXPECT_EQ(1, count);
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that MD5 hashes are computed by default.
-TEST_F(ObjectIntegrationTest, DefaultMD5HashJSON) {
-  Client client(ClientOptions()
-                    .set_enable_raw_client_tracing(true)
-                    .set_enable_http_tracing(true));
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  auto backend = std::make_shared<testing_util::CaptureLogLinesBackend>();
-  auto id = LogSink::Instance().AddBackend(backend);
-  StatusOr<ObjectMetadata> insert_meta = client.InsertObject(
-      bucket_name, object_name, LoremIpsum(), IfGenerationMatch(0));
-  ASSERT_TRUE(insert_meta.ok()) << "status=" << insert_meta.status();
-
-  LogSink::Instance().RemoveBackend(id);
-
-  auto count = std::count_if(
-      backend->log_lines.begin(), backend->log_lines.end(),
-      [](std::string const& line) {
-        // This is a big indirect, we detect if the upload changed to
-        // multipart/related, and if so, we assume the hash value is being used.
-        // Unfortunately I (@coryan) cannot think of a way to examine the upload
-        // contents.
-        return line.rfind("content-type: multipart/related; boundary=", 0) == 0;
-      });
-  EXPECT_EQ(1, count);
-
-  if (insert_meta->has_metadata("x_testbench_upload")) {
-    // When running against the testbench, we have some more information to
-    // verify the right upload type and contents were sent.
-    EXPECT_EQ("multipart", insert_meta->metadata("x_testbench_upload"));
-    ASSERT_TRUE(insert_meta->has_metadata("x_testbench_md5"));
-    auto expected_md5 = ComputeMD5Hash(LoremIpsum());
-    EXPECT_EQ(expected_md5, insert_meta->metadata("x_testbench_md5"));
-  }
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that `DisableMD5Hash` actually disables the header.
-TEST_F(ObjectIntegrationTest, DisableMD5HashXML) {
-  Client client(ClientOptions()
-                    .set_enable_raw_client_tracing(true)
-                    .set_enable_http_tracing(true));
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  auto backend = std::make_shared<testing_util::CaptureLogLinesBackend>();
-  auto id = LogSink::Instance().AddBackend(backend);
-  StatusOr<ObjectMetadata> insert_meta = client.InsertObject(
-      bucket_name, object_name, LoremIpsum(), IfGenerationMatch(0),
-      DisableMD5Hash(true), Fields(""));
-  ASSERT_TRUE(insert_meta.ok()) << "status=" << insert_meta.status();
-
-  LogSink::Instance().RemoveBackend(id);
-
-  auto count =
-      std::count_if(backend->log_lines.begin(), backend->log_lines.end(),
-                    [](std::string const& line) {
-                      return line.rfind("x-goog-hash: md5=", 0) == 0;
-                    });
-  EXPECT_EQ(0U, count);
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that `DisableMD5Hash` actually disables the payload.
-TEST_F(ObjectIntegrationTest, DisableMD5HashJSON) {
-  Client client(ClientOptions()
-                    .set_enable_raw_client_tracing(true)
-                    .set_enable_http_tracing(true));
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  auto backend = std::make_shared<testing_util::CaptureLogLinesBackend>();
-  auto id = LogSink::Instance().AddBackend(backend);
-  StatusOr<ObjectMetadata> insert_meta =
-      client.InsertObject(bucket_name, object_name, LoremIpsum(),
-                          IfGenerationMatch(0), DisableMD5Hash(true));
-  ASSERT_TRUE(insert_meta.ok()) << "status=" << insert_meta.status();
-
-  LogSink::Instance().RemoveBackend(id);
-
-  auto count = std::count_if(
-      backend->log_lines.begin(), backend->log_lines.end(),
-      [](std::string const& line) {
-        // This is a big indirect, we detect if the upload changed to
-        // multipart/related, and if so, we assume the hash value is being used.
-        // Unfortunately I (@coryan) cannot think of a way to examine the upload
-        // contents.
-        return line.rfind("content-type: multipart/related; boundary=", 0) == 0;
-      });
-  EXPECT_EQ(0, count);
-
-  if (insert_meta->has_metadata("x_testbench_upload")) {
-    // When running against the testbench, we have some more information to
-    // verify the right upload type and contents were sent.
-    EXPECT_EQ("simple", insert_meta->metadata("x_testbench_upload"));
-    ASSERT_FALSE(insert_meta->has_metadata("x_testbench_md5"));
-  }
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that MD5 hashes are computed by default on downloads.
-TEST_F(ObjectIntegrationTest, DefaultMD5StreamingReadXML) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  // Create an object and a stream to read it back.
-  StatusOr<ObjectMetadata> meta =
-      client.InsertObject(bucket_name, object_name, LoremIpsum(),
-                          IfGenerationMatch(0), Projection::Full());
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-
-  auto stream = client.ReadObject(bucket_name, object_name);
-  std::string actual(std::istreambuf_iterator<char>{stream}, {});
-  ASSERT_FALSE(stream.IsOpen());
-  ASSERT_FALSE(actual.empty());
-
-  EXPECT_EQ(stream.received_hash(), stream.computed_hash());
-  EXPECT_THAT(stream.received_hash(), HasSubstr(meta->md5_hash()));
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that MD5 hashes are computed by default on downloads.
-TEST_F(ObjectIntegrationTest, DefaultMD5StreamingReadJSON) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  // Create an object and a stream to read it back.
-  StatusOr<ObjectMetadata> meta =
-      client.InsertObject(bucket_name, object_name, LoremIpsum(),
-                          IfGenerationMatch(0), Projection::Full());
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-
-  auto stream =
-      client.ReadObject(bucket_name, object_name, IfMetagenerationNotMatch(0));
-  std::string actual(std::istreambuf_iterator<char>{stream}, {});
-  ASSERT_FALSE(stream.IsOpen());
-  ASSERT_FALSE(actual.empty());
-
-  EXPECT_EQ(stream.received_hash(), stream.computed_hash());
-  EXPECT_THAT(stream.received_hash(), HasSubstr(meta->md5_hash()));
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that hashes and checksums can be disabled on downloads.
-TEST_F(ObjectIntegrationTest, DisableHashesStreamingReadXML) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  // Create an object and a stream to read it back.
-  StatusOr<ObjectMetadata> meta =
-      client.InsertObject(bucket_name, object_name, LoremIpsum(),
-                          IfGenerationMatch(0), Projection::Full());
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-
-  auto stream =
-      client.ReadObject(bucket_name, object_name, DisableMD5Hash(true),
-                        DisableCrc32cChecksum(true));
-  std::string actual(std::istreambuf_iterator<char>{stream}, {});
-  ASSERT_FALSE(stream.IsOpen());
-  ASSERT_FALSE(actual.empty());
-
-  EXPECT_TRUE(stream.computed_hash().empty());
-  EXPECT_TRUE(stream.received_hash().empty());
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that hashes and checksums can be disabled on downloads.
-TEST_F(ObjectIntegrationTest, DisableHashesStreamingReadJSON) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  // Create an object and a stream to read it back.
-  StatusOr<ObjectMetadata> meta =
-      client.InsertObject(bucket_name, object_name, LoremIpsum(),
-                          IfGenerationMatch(0), Projection::Full());
-  ASSERT_TRUE(meta.ok()) << "status=" << meta.status();
-
-  auto stream = client.ReadObject(
-      bucket_name, object_name, DisableMD5Hash(true),
-      DisableCrc32cChecksum(true), IfMetagenerationNotMatch(0));
-  std::string actual(std::istreambuf_iterator<char>{stream}, {});
-  ASSERT_FALSE(stream.IsOpen());
-  ASSERT_FALSE(actual.empty());
-
-  EXPECT_TRUE(stream.computed_hash().empty());
-  EXPECT_TRUE(stream.received_hash().empty());
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that MD5 hashes are computed by default on uploads.
-TEST_F(ObjectIntegrationTest, DefaultMD5StreamingWriteXML) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  // Create the object, but only if it does not exist already.
-  auto os = client.WriteObject(bucket_name, object_name, IfGenerationMatch(0),
-                               Fields(""));
-  os.exceptions(std::ios_base::failbit);
-  // We will construct the expected response while streaming the data up.
-  std::ostringstream expected;
-  WriteRandomLines(os, expected);
-
-  auto expected_md5hash = ComputeMD5Hash(expected.str());
-
-  os.Close();
-  ObjectMetadata meta = os.metadata().value();
-  EXPECT_EQ(os.received_hash(), os.computed_hash());
-  EXPECT_THAT(os.received_hash(), HasSubstr(expected_md5hash));
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that MD5 hashes are computed by default on uploads.
-TEST_F(ObjectIntegrationTest, DefaultMD5StreamingWriteJSON) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  // Create the object, but only if it does not exist already.
-  auto os = client.WriteObject(bucket_name, object_name, IfGenerationMatch(0));
-  os.exceptions(std::ios_base::failbit);
-  // We will construct the expected response while streaming the data up.
-  std::ostringstream expected;
-  WriteRandomLines(os, expected);
-
-  auto expected_md5hash = ComputeMD5Hash(expected.str());
-
-  os.Close();
-  ObjectMetadata meta = os.metadata().value();
-  EXPECT_EQ(os.received_hash(), os.computed_hash());
-  EXPECT_THAT(os.received_hash(), HasSubstr(expected_md5hash));
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that hashes and checksums can be disabled in uploads.
-TEST_F(ObjectIntegrationTest, DisableHashesStreamingWriteXML) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  // Create the object, but only if it does not exist already.
-  auto os = client.WriteObject(bucket_name, object_name, IfGenerationMatch(0),
-                               Fields(""), DisableMD5Hash(true),
-                               DisableCrc32cChecksum(true));
-  os.exceptions(std::ios_base::failbit);
-  // We will construct the expected response while streaming the data up.
-  std::ostringstream expected;
-  WriteRandomLines(os, expected);
-
-  os.Close();
-  ObjectMetadata meta = os.metadata().value();
-  EXPECT_TRUE(os.received_hash().empty());
-  EXPECT_TRUE(os.computed_hash().empty());
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-/// @test Verify that hashes and checksums can be disabled in uploads.
-TEST_F(ObjectIntegrationTest, DisableHashesStreamingWriteJSON) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-
-  // Create the object, but only if it does not exist already.
-  auto os =
-      client.WriteObject(bucket_name, object_name, IfGenerationMatch(0),
-                         DisableMD5Hash(true), DisableCrc32cChecksum(true));
-  os.exceptions(std::ios_base::failbit);
-  // We will construct the expected response while streaming the data up.
-  std::ostringstream expected;
-  WriteRandomLines(os, expected);
-
-  os.Close();
-  ObjectMetadata meta = os.metadata().value();
-  EXPECT_TRUE(os.received_hash().empty());
-  EXPECT_TRUE(os.computed_hash().empty());
-
-  StatusOr<void> status = client.DeleteObject(bucket_name, object_name);
-  ASSERT_TRUE(status.ok()) << "status=" << status.status();
-}
-
-TEST_F(ObjectIntegrationTest, CopyFailure) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto source_object_name = MakeRandomObjectName();
-  auto destination_object_name = MakeRandomObjectName();
-
-  // This operation should fail because the source object does not exist.
-  auto meta = client.CopyObject(bucket_name, source_object_name, bucket_name,
-                      destination_object_name);
-  EXPECT_FALSE(meta.ok()) << "value=" << meta.value();
-}
-
 TEST_F(ObjectIntegrationTest, GetObjectMetadataFailure) {
   Client client;
   auto bucket_name = ObjectTestEnvironment::bucket_name();
@@ -1414,32 +694,6 @@ TEST_F(ObjectIntegrationTest, PatchObjectFailure) {
   auto patch = client.PatchObject(bucket_name, object_name,
                                   ObjectMetadataPatchBuilder());
   EXPECT_FALSE(patch.ok()) << "value=" << patch.value();
-}
-
-TEST_F(ObjectIntegrationTest, ComposeFailure) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto object_name = MakeRandomObjectName();
-  auto composed_object_name = MakeRandomObjectName();
-  std::vector<ComposeSourceObject> source_objects = {{object_name},
-                                                     {object_name}};
-
-  // This operation should fail because the source object does not exist.
-  auto meta =
-      client.ComposeObject(bucket_name, source_objects, composed_object_name);
-  EXPECT_FALSE(meta.ok()) << "value=" << meta.value();
-}
-
-TEST_F(ObjectIntegrationTest, RewriteFailure) {
-  Client client;
-  auto bucket_name = ObjectTestEnvironment::bucket_name();
-  auto source_object_name = MakeRandomObjectName();
-  auto destination_object_name = MakeRandomObjectName();
-
-  // This operation should fail because the source object does not exist.
-  StatusOr<ObjectMetadata> metadata = client.RewriteObjectBlocking(
-      bucket_name, source_object_name, bucket_name, destination_object_name);
-  EXPECT_FALSE(metadata.ok());
 }
 
 TEST_F(ObjectIntegrationTest, ListAccessControlFailure) {
